move tab to class file
