+++ nav-app/src/app/classes/index.ts
@@ -8,3 +8,4 @@ export { Link } from "./link";
 export { Metadata } from "./metadata";
 export { TechniqueVM } from "./technique-vm";
 export { ViewModel } from "./view-model";
+export { Tab } from "./tab";
