+++ nav-app/src/app/tabs/tabs.component.html
@@ -34,7 +34,7 @@
            *ngFor="let tab of layerTabs; let $i = index"
            (click)="handleTabClick(tab)"
            [active]="activeTab === tab">
-            {{tab.isDataTable ? tab.dataContext.name : tab.title}}
+            {{tab.isDataTable ? tab.viewModel.name : tab.title}}
             <span class="tab-close" *ngIf="tab.isCloseable" (click)="closeTab(tab)">
                 <svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 0 24 24" width="24px" [attr.fill]="activeTab === tab ? '#000000' : '#9aa0a6'"><path d="M0 0h24v24H0z" fill="none"/><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>
             </span>
@@ -541,5 +541,5 @@
 </ng-template>
 
 <ng-template [ngIf]="activeTab && activeTab.isDataTable">
-    <DataTable [viewModel]="activeTab.dataContext" [currentDropdown]="dropdownEnabled" (dropdownChange)="dropdownEnabled = $event" (onScroll)="adjustHeader($event)"></DataTable>
+    <DataTable [viewModel]="activeTab.viewModel" [currentDropdown]="dropdownEnabled" (dropdownChange)="dropdownEnabled = $event" (onScroll)="adjustHeader($event)"></DataTable>
 </ng-template>
